[Edited] - Edit Code Structure
