+++ src/App.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import './App.css';
 //import LandingPage from './components/LandingPage/LandingPage';
-import Navbar from './components/LandingPage/Navbar'
-import StoreInfo from "./components/store-detail/StoreInfo";
-import PaymentPage from "./pages/PaymentPage";
+// import Navbar from './components/landing/Navbar'
+// import StoreInfo from "./components/store-detail/StoreInfo";
+// import PaymentPage from "./pages/PaymentPage";
 
 
 function App() {
