--- src/components/LandingPage/LandingPage.js
@@ -1,15 +1,0 @@
-// import React, { Component } from 'react'
-// import PageInfo from './/PageInfo'
-// import Navbar from './/Navbar'
-
-// export class LandingPage extends Component {
-//     render() {
-//         return (
-//             <div>
-//                 <PageInfo></PageInfo>
-//             </div>
-//         )
-//     }
-// }
-
-// export default LandingPage
