--- src/components/LandingPage/SearchZone/SearchInput.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-export class SearchInput extends Component {
-    render() {
-        return (
-            <div>
-                <span>Search Here!!</span>
-            </div>
-        )
-    }
-}
-
-export default SearchInput
