--- src/components/LandingPage/SearchZone/SearchResult.js
@@ -1,13 +1,0 @@
-import React, { Component } from 'react'
-
-export class SearchResult extends Component {
-    render() {
-        return (
-            <div>
-                <span>Store Here!!</span>
-            </div>
-        )
-    }
-}
-
-export default SearchResult
