+++ src/components/landing/search-zone/SearchInput.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react'
 
-class SearchInput extends Component {
-  render() {
-    return (
-      <div>
-        
-      </div>
-    )
-  }
+export class SearchInput extends Component {
+    render() {
+        return (
+            <div>
+                <span>Search Here!!</span>
+            </div>
+        )
+    }
 }
 
 export default SearchInput
