+++ src/components/landing/search-zone/SearchResult.js
@@ -1,13 +1,13 @@
 import React, { Component } from 'react'
 
-class SearchResult extends Component {
-  render() {
-    return (
-      <div>
-        
-      </div>
-    )
-  }
+export class SearchResult extends Component {
+    render() {
+        return (
+            <div>
+                <span>Store Here!!</span>
+            </div>
+        )
+    }
 }
 
 export default SearchResult
