+++ src/pages/LandingPage.js
@@ -1,13 +1,15 @@
 import React, { Component } from 'react'
+import PageInfo from '../components/landing/PageInfo'
+import Navbar from '../components/landing/Navbar'
 
-class LandingPage extends Component {
-  render() {
-    return (
-      <div>
-        
-      </div>
-    )
-  }
+export class LandingPage extends Component {
+    render() {
+        return (
+            <div>
+                <PageInfo></PageInfo>
+            </div>
+        )
+    }
 }
 
 export default LandingPage
